// src/types/issueReport.types.ts
export enum EnumPriority {
    CRITICAL = 'CRITICAL',
    HIGH = 'HIGH',
    LOW = 'LOW',
    NORMAL = 'NORMAL'
}

export enum EnumStatus {
    CLOSED = 'CLOSED',
    DISCUSSION = 'DISCUSSION',
    FULFILMENT = 'FULFILMENT',
    READY = 'READY',
    TO_APPROVE = 'TO_APPROVE'
}

export enum EnumType {
    FUNCTIONALITY = 'FUNCTIONALITY',
    REPORTING = 'REPORTING',
    UI = 'UI',
    VULNERABILITY = 'VULNERABILITY'
}

export enum EnumRole {
    DEV = 'DEV',
    GUEST = 'GUEST',
    QA = 'QA'
}